+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 - Cloud storage: fix potential issue with credentials on Linux [#1346]
 - Mobile/iOS: fix missing translations
+- Dive media: sort thumbnails by timestamp
+- Dive media: don't recalculate all pictures on drag & drop
 - Dive media: support addition of videos
 - Dive media: locate moved files based on filename and path
+- Profile: immediately update thumbnail positions on deletion
 - Profile: Context menu entry to manually split a dive
 - BLE support: fix recognition of Heinrich Weikamp OSTC 2 and OSTC Plus
 ---